+++ src/test/java/roomescape/service/reservation/ReservationServiceTest.java
@@ -1,4 +1,4 @@
-package roomescape.service;
+package roomescape.service.reservation;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -9,20 +9,15 @@ import roomescape.domain.Member;
 import roomescape.domain.Password;
 import roomescape.domain.ReservationStatus;
 import roomescape.domain.Role;
-import roomescape.domain.dto.ReservationRequest;
 import roomescape.domain.dto.ReservationResponse;
 import roomescape.domain.dto.ReservationWaitingResponse;
 import roomescape.domain.dto.ReservationsMineResponse;
 import roomescape.domain.dto.ResponsesWrapper;
-import roomescape.exception.InvalidClientFieldWithValueException;
-import roomescape.exception.ReservationFailException;
 import roomescape.repository.ReservationRepository;
 
 import java.time.LocalDate;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertAll;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
@@ -37,7 +32,7 @@ class ReservationServiceTest {
     }
 
     private long getReservationSize() {
-        return service.findEntireReservations().getData().size();
+        return repository.findAll().size();
     }
 
     @Test
@@ -46,21 +41,6 @@ class ReservationServiceTest {
         assertThat(service.findEntireReservations().getData().size()).isEqualTo(10);
     }
 
-    @Test
-    @DisplayName("예약이 성공하면 결과값과 함께 Db에 저장된다.")
-    void given_reservationRequestWithInitialSize_when_register_then_returnReservationResponseAndSaveDb() {
-        //given
-        long initialSize = getReservationSize();
-        final ReservationRequest reservationRequest = new ReservationRequest(LocalDate.parse("2999-01-01"), 1L, 1L, 1L);
-        //when
-        final ReservationResponse reservationResponse = service.register(reservationRequest);
-        long afterCreateSize = getReservationSize();
-        //then
-        assertAll(
-                () -> assertThat(reservationResponse.id()).isEqualTo(afterCreateSize),
-                () -> assertThat(afterCreateSize).isEqualTo(initialSize + 1)
-        );
-    }
 
     @Test
     @DisplayName("존재하는 예약을 삭제하면 Db에도 삭제된다.")
@@ -72,64 +52,6 @@ class ReservationServiceTest {
         long afterCreateSize = getReservationSize();
         //then
         assertThat(afterCreateSize).isEqualTo(initialSize - 1);
-    }
-
-    @Test
-    @DisplayName("이전 날짜로 예약 할 경우 예외가 발생하고, Db에 저장하지 않는다.")
-    void given_reservationRequestWithInitialSize_when_registerWithPastDate_then_throwException() {
-        //given
-        long initialSize = getReservationSize();
-        final ReservationRequest reservationRequest = new ReservationRequest(LocalDate.parse("1999-01-01"), 1L, 1L, 1L);
-        //when, then
-        assertAll(
-                () -> assertThatThrownBy(() -> service.register(reservationRequest)).isInstanceOf(ReservationFailException.class),
-                () -> assertThat(getReservationSize()).isEqualTo(initialSize)
-        );
-    }
-
-    @Test
-    @DisplayName("themeId가 존재하지 않을 경우 예외를 발생하고, Db에 저장하지 않는다.")
-    void given_reservationRequestWithInitialSize_when_registerWithNotExistThemeId_then_throwException() {
-        //given
-        long initialSize = getReservationSize();
-        final ReservationRequest reservationRequest = new ReservationRequest(LocalDate.parse("2099-01-01"), 1L, 99L,
-                1L);
-        //when, then
-        assertAll(
-                () -> assertThatThrownBy(() -> service.register(reservationRequest)).isInstanceOf(
-                        InvalidClientFieldWithValueException.class),
-                () -> assertThat(getReservationSize()).isEqualTo(initialSize)
-        );
-    }
-
-    @Test
-    @DisplayName("timeId 존재하지 않을 경우 예외를 발생하고, Db에 저장하지 않는다.")
-    void given_reservationRequestWithInitialSize_when_registerWithNotExistTimeId_then_throwException() {
-        //given
-        long initialSize = getReservationSize();
-        final ReservationRequest reservationRequest = new ReservationRequest(LocalDate.parse("2099-01-01"), 99L, 1L,
-                1L);
-        //when, then
-        assertAll(
-                () -> assertThatThrownBy(() -> service.register(reservationRequest)).isInstanceOf(
-                        InvalidClientFieldWithValueException.class),
-                () -> assertThat(getReservationSize()).isEqualTo(initialSize)
-        );
-    }
-
-    @Test
-    @DisplayName("memberId 존재하지 않을 경우 예외를 발생하고, Db에 저장하지 않는다.")
-    void given_reservationRequestWithInitialSize_when_registerWithNotExistMemberId_then_throwException() {
-        //given
-        long initialSize = getReservationSize();
-        final ReservationRequest reservationRequest = new ReservationRequest(LocalDate.parse("2099-01-01"), 1L, 1L,
-                99L);
-        //when, then
-        assertAll(
-                () -> assertThatThrownBy(() -> service.register(reservationRequest)).isInstanceOf(
-                        InvalidClientFieldWithValueException.class),
-                () -> assertThat(getReservationSize()).isEqualTo(initialSize)
-        );
     }
 
     @Test
@@ -167,22 +89,6 @@ class ReservationServiceTest {
         final ResponsesWrapper<ReservationResponse> reservationResponses = service.findReservations(themeId, memberId, dateFrom, dateTo);
         //then
         assertThat(reservationResponses.getData()).hasSize(3);
-    }
-
-    @Test
-    @DisplayName("이미 예약이 되어있는 날짜와 시간 및 테마에 다른 사용자가 예약 등록을 할 경우 예약이 저장된다.")
-    void given_reservationRequest_when_registerAlreadyReservedWithDifferentMemberId_then_createdWithStatusIsWaiting() {
-        //given
-        long initialSize = getReservationSize();
-        final ReservationRequest reservationRequest = new ReservationRequest(LocalDate.parse("2099-04-30"), 1L, 1L, 2L);
-        //when
-        final ReservationResponse reservationResponse = service.register(reservationRequest);
-        long afterCreateSize = getReservationSize();
-        //then
-        assertAll(
-                () -> assertThat(reservationResponse.id()).isEqualTo(afterCreateSize),
-                () -> assertThat(afterCreateSize).isEqualTo(initialSize + 1)
-        );
     }
 
     @Test
